Refactor PGException and QgsPostgresResult
